Add files via upload
The PA_new.py is a class for porfiolio analysis. It supports not only the univarable analysis, but also the bivarable analysis for which the depent and indepent manner are both admitted. To accomplish more utilies, the transition matrix and cal the alpha of factors model can also be used. The analyse.py is a demo for use the PA_new. The result_plt.py is to visualize the bivarable analysis result by 3d bar grapha.